Moved config to stream_config to clarify the purpose of those config typs
