+++ sdk/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod args;
 pub mod binary;
-pub mod builder;
 pub mod bytes_serializable;
 #[cfg(feature = "iggy-cli")]
 pub mod cli;
@@ -14,7 +13,6 @@ pub mod client_provider;
 pub mod clients;
 pub mod command;
 pub mod compression;
-pub mod config;
 pub mod confirmation;
 pub mod consumer;
 pub mod consumer_ext;
@@ -32,6 +30,8 @@ pub mod partitions;
 pub mod personal_access_tokens;
 pub mod quic;
 pub mod snapshot;
+pub mod stream_builder;
+pub mod stream_config;
 pub mod streams;
 pub mod system;
 pub mod tcp;
