+++ sdk/src/stream_config/config_iggy_producer.rs
@@ -1,31 +1,26 @@
 use bon::Builder;
 
-use crate::clients::consumer::{AutoCommit, AutoCommitWhen};
-use crate::config::shared_config;
-use crate::consumer::ConsumerKind;
 use crate::identifier::Identifier;
-use crate::messages::poll_messages::PollingStrategy;
+use crate::messages::send_messages::Partitioning;
+use crate::stream_config::shared_config;
 use crate::utils::duration::IggyDuration;
 use std::str::FromStr;
 
 #[derive(Builder, Debug, Clone)]
 #[builder(on(String, into))]
-pub struct IggyConsumerConfig {
+pub struct IggyProducerConfig {
     stream_id: Identifier,
     stream_name: String,
     topic_id: Identifier,
     topic_name: String,
-    auto_commit: AutoCommit,
     batch_size: u32,
-    consumer_name: String,
-    consumer_kind: ConsumerKind,
-    polling_interval: IggyDuration,
-    polling_strategy: PollingStrategy,
+    send_interval: IggyDuration,
+    partitioning: Partitioning,
     partitions_count: u32,
     replication_factor: Option<u8>,
 }
 
-impl Default for IggyConsumerConfig {
+impl Default for IggyProducerConfig {
     fn default() -> Self {
         let stream_id = shared_config::get_identifier_from_string("test_stream");
         let topic_id = shared_config::get_identifier_from_string("test_topic");
@@ -35,20 +30,17 @@ impl Default for IggyConsumerConfig {
             stream_name: "test_stream".to_string(),
             topic_id,
             topic_name: "test_topic".to_string(),
-            auto_commit: AutoCommit::When(AutoCommitWhen::PollingMessages),
             batch_size: 100,
-            consumer_name: "test_consumer".to_string(),
-            consumer_kind: ConsumerKind::ConsumerGroup,
-            polling_interval: IggyDuration::from_str("5ms").unwrap(),
-            polling_strategy: PollingStrategy::last(),
+            send_interval: IggyDuration::from_str("5ms").unwrap(),
+            partitioning: Partitioning::balanced(),
             partitions_count: 1,
             replication_factor: None,
         }
     }
 }
 
-impl IggyConsumerConfig {
-    /// Creates a new `IggyConsumerConfig` from the given arguments.
+impl IggyProducerConfig {
+    /// Creates a new `IggyProducerConfig` with all fields defined.
     ///
     /// # Args
     ///
@@ -56,28 +48,23 @@ impl IggyConsumerConfig {
     /// * `stream_name` - The stream name.
     /// * `topic_id` - The topic identifier.
     /// * `topic_name` - The topic name.
-    /// * `auto_commit` - The auto-commit configuration to use.
     /// * `batch_size` - The max number of messages to send in a batch.
-    /// * `consumer_name` - The name of the consumer group.
-    /// * `consumer_kind` - The consumer kind to use.
-    /// * `polling_interval` - The interval between polling for new messages.
-    /// * `polling_strategy` - The polling strategy to use.
+    /// * `send_interval` - The interval between messages sent.
+    /// * `partitioning` - The partitioning strategy to use.
     /// * `partition` - The number of partitions to create.
+    /// * `replication_factor` - The replication factor to use.
     ///
     /// Returns:
-    /// A new `IggyConsumerConfig`.
+    /// A new `IggyProducerConfig`.
     ///
     pub fn new(
         stream_id: Identifier,
         stream_name: String,
         topic_id: Identifier,
         topic_name: String,
-        auto_commit: AutoCommit,
         batch_size: u32,
-        consumer_name: String,
-        consumer_kind: ConsumerKind,
-        polling_interval: IggyDuration,
-        polling_strategy: PollingStrategy,
+        send_interval: IggyDuration,
+        partitioning: Partitioning,
         partitions_count: u32,
         replication_factor: Option<u8>,
     ) -> Self {
@@ -86,34 +73,32 @@ impl IggyConsumerConfig {
             stream_name,
             topic_id,
             topic_name,
-            auto_commit,
             batch_size,
-            consumer_name,
-            consumer_kind,
-            polling_interval,
-            polling_strategy,
+            send_interval,
+            partitioning,
             partitions_count,
             replication_factor,
         }
     }
 
-    /// Creates a new `IggyConsumerConfig` from the given arguments.
+    /// Creates a new `IggyProducerConfig` from the given stream and topic names, along with the
+    /// max batch size and the send interval.
     ///
     /// # Args
     ///
     /// * `stream` - The stream name.
     /// * `topic` - The topic name.
     /// * `batch_size` - The max number of messages to send in a batch.
-    /// * `polling_interval` - The interval between polling for new messages.
+    /// * `send_interval` - The interval between messages sent.
     ///
     /// Returns:
-    /// A new `IggyConsumerConfig`.
+    /// A new `IggyProducerConfig`.
     ///
     pub fn from_stream_topic(
         stream: &str,
         topic: &str,
         batch_size: u32,
-        polling_interval: IggyDuration,
+        send_interval: IggyDuration,
     ) -> Self {
         let stream_id = shared_config::get_identifier_from_string(stream);
         let topic_id = shared_config::get_identifier_from_string(topic);
@@ -123,19 +108,16 @@ impl IggyConsumerConfig {
             stream_name: stream.to_string(),
             topic_id,
             topic_name: topic.to_string(),
-            auto_commit: AutoCommit::When(AutoCommitWhen::PollingMessages),
             batch_size,
-            consumer_name: format!("consumer-{}-{}", stream, topic),
-            consumer_kind: ConsumerKind::ConsumerGroup,
-            polling_interval,
-            polling_strategy: PollingStrategy::last(),
+            send_interval,
+            partitioning: Partitioning::balanced(),
             partitions_count: 1,
             replication_factor: None,
         }
     }
 }
 
-impl IggyConsumerConfig {
+impl IggyProducerConfig {
     pub fn stream_id(&self) -> &Identifier {
         &self.stream_id
     }
@@ -152,28 +134,16 @@ impl IggyConsumerConfig {
         &self.topic_name
     }
 
-    pub fn auto_commit(&self) -> AutoCommit {
-        self.auto_commit
-    }
-
     pub fn batch_size(&self) -> u32 {
         self.batch_size
     }
 
-    pub fn consumer_name(&self) -> &str {
-        &self.consumer_name
+    pub fn send_interval(&self) -> IggyDuration {
+        self.send_interval
     }
 
-    pub fn consumer_kind(&self) -> ConsumerKind {
-        self.consumer_kind
-    }
-
-    pub fn polling_interval(&self) -> IggyDuration {
-        self.polling_interval
-    }
-
-    pub fn polling_strategy(&self) -> PollingStrategy {
-        self.polling_strategy
+    pub fn partitioning(&self) -> &Partitioning {
+        &self.partitioning
     }
 
     pub fn partitions_count(&self) -> u32 {
@@ -192,17 +162,14 @@ mod tests {
     #[test]
     fn test_builder() {
         // Builder is generated by the bon macro
-        let config = IggyConsumerConfig::builder()
+        let config = IggyProducerConfig::builder()
             .stream_id(shared_config::get_identifier_from_string("test_stream"))
-            .stream_name("test_stream".to_string())
+            .stream_name("test_stream")
             .topic_id(shared_config::get_identifier_from_string("test_topic"))
-            .topic_name("test_topic".to_string())
-            .auto_commit(AutoCommit::When(AutoCommitWhen::PollingMessages))
+            .topic_name("test_topic")
             .batch_size(100)
-            .consumer_name("test_consumer".to_string())
-            .consumer_kind(ConsumerKind::ConsumerGroup)
-            .polling_interval(IggyDuration::from_str("5ms").unwrap())
-            .polling_strategy(PollingStrategy::last())
+            .send_interval(IggyDuration::from_str("5ms").unwrap())
+            .partitioning(Partitioning::balanced())
             .partitions_count(1)
             .build();
 
@@ -216,19 +183,14 @@ mod tests {
             &shared_config::get_identifier_from_string("test_topic")
         );
         assert_eq!(config.topic_name(), "test_topic");
-        assert_eq!(
-            config.auto_commit(),
-            AutoCommit::When(AutoCommitWhen::PollingMessages)
-        );
         assert_eq!(config.batch_size(), 100);
-        assert_eq!(config.consumer_name(), "test_consumer");
-        assert_eq!(config.consumer_kind(), ConsumerKind::ConsumerGroup);
         assert_eq!(
-            config.polling_interval(),
+            config.send_interval(),
             IggyDuration::from_str("5ms").unwrap()
         );
-        assert_eq!(config.polling_strategy(), PollingStrategy::last());
+        assert_eq!(config.partitioning(), &Partitioning::balanced());
         assert_eq!(config.partitions_count(), 1);
+        assert_eq!(config.replication_factor(), None);
     }
 
     #[test]
@@ -236,87 +198,72 @@ mod tests {
         let stream_id = shared_config::get_identifier_from_string("test_stream");
         let topic_id = shared_config::get_identifier_from_string("test_topic");
 
-        let config = IggyConsumerConfig::default();
+        let config = IggyProducerConfig::default();
         assert_eq!(config.stream_id(), &stream_id);
         assert_eq!(config.stream_name(), "test_stream");
         assert_eq!(config.topic_id(), &topic_id);
         assert_eq!(config.topic_name(), "test_topic");
-        assert_eq!(
-            config.auto_commit(),
-            AutoCommit::When(AutoCommitWhen::PollingMessages)
-        );
         assert_eq!(config.batch_size(), 100);
-        assert_eq!(config.consumer_name(), "test_consumer");
-        assert_eq!(config.consumer_kind(), ConsumerKind::ConsumerGroup);
         assert_eq!(
-            config.polling_interval(),
+            config.send_interval(),
             IggyDuration::from_str("5ms").unwrap()
         );
-        assert_eq!(config.polling_strategy(), PollingStrategy::last());
+        assert_eq!(config.partitioning(), &Partitioning::balanced());
         assert_eq!(config.partitions_count(), 1);
         assert_eq!(config.replication_factor(), None);
     }
 
     #[test]
     fn test_new() {
-        let config = IggyConsumerConfig::new(
-            shared_config::get_identifier_from_string("test_stream"),
-            "test_stream".to_string(),
-            shared_config::get_identifier_from_string("test_topic"),
-            "test_topic".to_string(),
-            AutoCommit::When(AutoCommitWhen::PollingMessages),
+        let stream_id = shared_config::get_identifier_from_string("test_stream");
+        let topic_id = shared_config::get_identifier_from_string("test_topic");
+
+        let config = IggyProducerConfig::new(
+            stream_id.clone(),
+            String::from("test_stream"),
+            topic_id.clone(),
+            String::from("test_topic"),
             100,
-            "test_consumer".to_string(),
-            ConsumerKind::ConsumerGroup,
             IggyDuration::from_str("5ms").unwrap(),
-            PollingStrategy::last(),
+            Partitioning::balanced(),
             1,
             None,
         );
-        assert_eq!(
-            config.stream_id(),
-            &shared_config::get_identifier_from_string("test_stream")
-        );
+        assert_eq!(config.stream_id(), &stream_id);
         assert_eq!(config.stream_name(), "test_stream");
-        assert_eq!(
-            config.topic_id(),
-            &shared_config::get_identifier_from_string("test_topic")
-        );
+        assert_eq!(config.topic_id(), &topic_id);
         assert_eq!(config.topic_name(), "test_topic");
-        assert_eq!(
-            config.auto_commit(),
-            AutoCommit::When(AutoCommitWhen::PollingMessages)
-        );
         assert_eq!(config.batch_size(), 100);
-        assert_eq!(config.consumer_name(), "test_consumer");
-        assert_eq!(config.consumer_kind(), ConsumerKind::ConsumerGroup);
         assert_eq!(
-            config.polling_interval(),
+            config.send_interval(),
             IggyDuration::from_str("5ms").unwrap()
         );
-        assert_eq!(config.polling_strategy(), PollingStrategy::last());
+        assert_eq!(config.partitioning(), &Partitioning::balanced());
         assert_eq!(config.partitions_count(), 1);
         assert_eq!(config.replication_factor(), None);
     }
 
     #[test]
     fn test_from_stream_topic() {
-        let config = IggyConsumerConfig::from_stream_topic(
+        let stream_id = shared_config::get_identifier_from_string("test_stream");
+        let topic_id = shared_config::get_identifier_from_string("test_topic");
+
+        let config = IggyProducerConfig::from_stream_topic(
             "test_stream",
             "test_topic",
             100,
             IggyDuration::from_str("5ms").unwrap(),
         );
+        assert_eq!(config.stream_id(), &stream_id);
         assert_eq!(config.stream_name(), "test_stream");
+        assert_eq!(config.topic_id(), &topic_id);
         assert_eq!(config.topic_name(), "test_topic");
         assert_eq!(config.batch_size(), 100);
-        assert_eq!(config.consumer_name(), "consumer-test_stream-test_topic");
-        assert_eq!(config.consumer_kind(), ConsumerKind::ConsumerGroup);
         assert_eq!(
-            config.polling_interval(),
+            config.send_interval(),
             IggyDuration::from_str("5ms").unwrap()
         );
-        assert_eq!(config.polling_strategy(), PollingStrategy::last());
+        assert_eq!(config.partitioning(), &Partitioning::balanced());
         assert_eq!(config.partitions_count(), 1);
         assert_eq!(config.replication_factor(), None);
     }
